FFHRS-46: Allow JS execution in nextjs - Hardcode navigation is as it's required by JS - Create two different index files - one for Storybook and another for general use
