+++ src/components/general/Navigation/navigation.html.twig
@@ -1,4 +1,4 @@
-<nav id="{{ id }}" class="navigation" aria-label="{{ label }}">
+<nav id="navigation" class="navigation" aria-label="{{ label }}">
   <div class="navigation__background-blur"></div>
   <div class="navigation__container">
     <div class="navigation__row">
